DAOS-2218 tests: Test to verify different value types to same keys
Case1: Insert akey,dkey with single value
       Insert same akey,dkey with array value
       Result: should return Error.
Case2: Insert akey,dkey with single value
       Punch the keys
       Trigger aggregation
       Insert same akey,dkey with array value
      Result: Should work without errors.
