+++ src/utils/py/daos_api.py
@@ -1496,18 +1496,6 @@ class DaosContainer(object):
         else:
             c_value = None
 
-        #if isinstance(thedata, int):
-        #    c_value = ctypes.c_int(thedata)
-        #elif thedata is not None:
-        #    c_value = ctypes.create_string_buffer(thedata)
-        #else:
-        #    c_value = None
-
-        #if size == "int":
-        #    c_size = ctypes.sizeof(ctypes.c_int(thedata))
-        #else:
-        #    c_size = ctypes.c_size_t(size)
-        
         c_size = ctypes.sizeof(c_value)
         if dkey is None:
             c_dkey = None
